Refactor CourseDetailSideBarItem
