+++ src/pages/TeacherPage/ManageCourses/TeacherCourseDetail/CourseDetailSideBarItem/CourseInformation/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./CourseInformation";
