+++ src/pages/TeacherPage/ManageCourses/TeacherCourseDetail/CourseDetailSideBarItem/IntendedLearners/IntendedLearners.scss
