+++ src/pages/TeacherPage/ManageCourses/TeacherCourseDetail/CourseDetailSideBarItem/LearningContent/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./LearningContent";
