+++ src/pages/TeacherPage/ManageCourses/TeacherCourseDetail/CourseDetailSideBarItem/Targets/Targets.scss
