+++ src/pages/TeacherPage/ManageCourses/TeacherCourseDetail/CourseDetailSideBarItem/index.tsx
@@ -1,0 +1,15 @@
+import CourseInformation from "./CourseInformation";
+import IntendedLearners from "./IntendedLearners";
+import LearningContent from "./LearningContent";
+import Requirements from "./Requirements";
+import Targets from "./Targets";
+
+const CourseDetailSideBarItem = {
+  CourseInformation: CourseInformation,
+  IntendedLearners: IntendedLearners,
+  Requirements: Requirements,
+  Targets: Targets,
+  LearningContent: LearningContent,
+};
+
+export default CourseDetailSideBarItem;
