+++ src/pages/TeacherPage/ManageCourses/TeacherCourseDetail/TeacherCourseDetail.tsx
@@ -1,39 +1,35 @@
 import { Button } from "@mui/material";
-import { Box } from "@mui/system";
 import { useFormik } from "formik";
 import React, { useCallback, useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
 import { useNavigate, useParams } from "react-router-dom";
 import { toast } from "react-toastify";
-import categoryApi from "src/apis/categoryApi";
 import chapterApi from "src/apis/chapterApi";
 import courseApi from "src/apis/courseApi";
 import teacherApi from "src/apis/teacherApi";
-import FormControl from "src/components/FormControl";
-import MediaContent from "src/components/MediaContent";
 import LayoutContainer from "src/layouts/LayoutContainer";
 import { isPending, isSuccess } from "src/reducers";
-import { CourseStatus } from "src/types";
-import * as Yup from "yup";
-import IntendedLearners from "../IntendedLearners";
-import Requirements from "../Requirements";
-import Targets from "../Targets";
-import Chapter, { IChapterUpload } from "./Chapter";
+import { IChapterUpload } from "./Chapter";
+import CourseDetailSideBar, {
+  CourseDetailSideBarData,
+} from "./CourseDetailSideBar";
+import CourseDetailSideBarItem from "./CourseDetailSideBarItem";
 import "./TeacherCourseDetail.scss";
-import { ICategoriesCourse } from "./TeacherCourseDetail.type";
 
 const TeacherCourseDetail: React.FC = () => {
   document.title = "Quản lý khóa học";
 
   const { id } = useParams();
   const dispatch = useDispatch();
+  const nav = useNavigate();
+
   const [navbar, setNavbar] = useState(0);
+  const [isChapterUpdated, setIsChapterUpdated] = useState(false);
+  const [isInformationCourseUpdated, setIsInformationCourseUpdated] =
+    useState(false);
   const [chapters, setChapters] = useState<IChapterUpload[]>([]);
-  const [categories, setCategories] = useState<ICategoriesCourse[]>([]);
   const [slug, setSlug] = useState("");
   const [courseStatus, setCourseStatus] = useState();
-  const [image, setImage] = useState("");
-  const nav = useNavigate();
 
   const formik = useFormik({
     enableReinitialize: true,
@@ -45,52 +41,8 @@ const TeacherCourseDetail: React.FC = () => {
       currentPrice: "",
       thumbnail: "",
     },
-    validationSchema: Yup.object({
-      name: Yup.string().required("Vui lòng nhập khóa học"),
-      originalPrice: Yup.number().required("Vui lòng nhập giá gốc khóa học"),
-      currentPrice: Yup.number().required(
-        "Vui lòng nhập giá khuyến mãi khóa học"
-      ),
-    }),
-    validate: (values) => {
-      let errors = {};
-      if (Number(values.currentPrice) > Number(values.originalPrice)) {
-        errors = {
-          ...errors,
-          currentPrice: "Giá khuyến mãi phải nhỏ hơn giá gốc",
-        };
-      }
-
-      return errors;
-    },
-    onSubmit: async (values) => {
-      dispatch(isPending());
-      courseApi
-        .updateCourse(slug, {
-          ...values,
-          thumbnail: image,
-          originalPrice: Number(values.originalPrice),
-          currentPrice: Number(values.currentPrice),
-        })
-        .then((res: any) => {
-          dispatch(isSuccess());
-          toast.success(res.message, {
-            position: "bottom-right",
-          });
-        });
-    },
+    onSubmit: () => {},
   });
-
-  useEffect(() => {
-    categoryApi.getCategories().then((res: any) => {
-      setCategories(
-        res.categories.map((category: any) => {
-          return { value: category._id, name: category.name };
-        })
-      );
-    });
-    // eslint-disable-next-line
-  }, []);
 
   const getCourseDetails = useCallback(() => {
     id &&
@@ -132,62 +84,26 @@ const TeacherCourseDetail: React.FC = () => {
   useEffect(() => {
     dispatch(isPending());
     getCourseDetails();
-  }, [getCourseDetails, dispatch]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
-  const handleAddChapter = (index: number) => {
-    dispatch(isPending());
-    chapterApi.addChapter(id, index, "default").then(() => {
+  useEffect(() => {
+    if (isInformationCourseUpdated) {
+      getCourseDetails();
+    }
+
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isInformationCourseUpdated]);
+
+  useEffect(() => {
+    if (isChapterUpdated) {
       chapterApi.getChapters(id).then((res: any) => {
         dispatch(isSuccess());
         setChapters(res.chapters);
       });
-    });
-  };
-
-  const postImage = (image: any) => {
-    dispatch(isPending());
-    const formData = new FormData();
-    formData.append("image", image);
-
-    courseApi
-      .uploadImage(formData)
-      .then((res: any) => {
-        dispatch(isSuccess());
-        const { message, url } = res;
-        toast.success(`${message}`, {
-          position: "bottom-right",
-        });
-        setImage(url);
-      })
-      .catch((err) => {
-        dispatch(isSuccess());
-        console.log("Lỗi rồi", err);
-      });
-  };
-
-  const handleUpdateChapter = (
-    name: string,
-    order: number,
-    chapterId: string
-  ) => {
-    dispatch(isPending());
-    chapterApi.updateChapter(chapterId, order, name).then(() => {
-      chapterApi.getChapters(id).then((res: any) => {
-        dispatch(isSuccess());
-        setChapters(res.chapters);
-      });
-    });
-  };
-
-  const handleDeleteChapter = (chapterId: string) => {
-    dispatch(isPending());
-    chapterApi.deleteChapter(chapterId).then(() => {
-      chapterApi.getChapters(id).then((res: any) => {
-        dispatch(isSuccess());
-        setChapters(res.chapters);
-      });
-    });
-  };
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [id, isChapterUpdated]);
 
   const handleSubmitReview = () => {
     dispatch(isPending());
@@ -225,38 +141,10 @@ const TeacherCourseDetail: React.FC = () => {
   return (
     <LayoutContainer titleShow={false} footerShow={false}>
       <div className="teacher-course-detail">
-        <div className="sidebar">
-          <h2>Danh sách các mục</h2>
-          <p
-            className={navbar === 0 ? "active" : ""}
-            onClick={() => setNavbar(0)}
-          >
-            Thông tin khóa học
-          </p>
-          <p
-            className={navbar === 1 ? "active" : ""}
-            onClick={() => setNavbar(1)}
-          >
-            Chương trình giảng dạy
-          </p>
-          <p
-            className={navbar === 2 ? "active" : ""}
-            onClick={() => setNavbar(2)}
-          >
-            Đối tượng khóa học
-          </p>
-          <p
-            className={navbar === 3 ? "active" : ""}
-            onClick={() => setNavbar(3)}
-          >
-            Kiến thức bắt buộc
-          </p>
-          <p
-            className={navbar === 4 ? "active" : ""}
-            onClick={() => setNavbar(4)}
-          >
-            Mục tiêu khóa học
-          </p>
+        <CourseDetailSideBar
+          sideBarContent={CourseDetailSideBarData}
+          getSideBarId={(sideBarId) => setNavbar(sideBarId as number)}
+        >
           <Button
             variant="contained"
             color="primary"
@@ -277,163 +165,38 @@ const TeacherCourseDetail: React.FC = () => {
           >
             Xóa khóa học
           </Button>
-        </div>
+          <Button
+            variant="contained"
+            color="inherit"
+            sx={{
+              height: 45,
+            }}
+            onClick={() => nav(-1)}
+          >
+            Quay lại trang trước
+          </Button>
+        </CourseDetailSideBar>
         <div className="form">
-          <h2 className="title">
-            {navbar === 0
-              ? "Thông tin khóa học"
-              : navbar === 1
-              ? "Chương trình giảng dạy"
-              : navbar === 2
-              ? "Đối tượng nào nên học?"
-              : navbar === 3
-              ? "Kiến thức bắt buộc cần có?"
-              : "Bạn sẽ học được gì?"}
-            {navbar === 0
-              ? courseStatus && (
-                  <p>
-                    Trạng thái khóa học:{" "}
-                    <span>{CourseStatus[courseStatus]}</span>
-                  </p>
-                )
-              : navbar === 1 && (
-                  <Button
-                    variant="contained"
-                    color="warning"
-                    sx={{
-                      height: 45,
-                    }}
-                    onClick={() => nav(`/teacher/course/preview-course/${id}`)}
-                  >
-                    Xem trước khóa học
-                  </Button>
-                )}
-          </h2>
           {navbar === 0 ? (
-            <form onSubmit={formik.handleSubmit}>
-              <Box
-                sx={{
-                  flex: 1,
-                  display: "flex",
-                  flexDirection: "column",
-                  gap: 1,
-                }}
-              >
-                <FormControl.InputUploadFile
-                  className="thumbnail-course"
-                  label="Thumbnail khóa học"
-                  valueDefault={formik.values.thumbnail}
-                  // onChange={(value) => formik.setFieldValue("thumbnail", value)}
-                  onChange={(value) => postImage(value)}
-                />
-                <FormControl.Input
-                  required
-                  label="Tên khóa học"
-                  placeholder="Nhập tên khóa học"
-                  errorMessage={formik.touched.name ? formik.errors.name : ""}
-                  {...formik.getFieldProps("name")}
-                />
-                <FormControl.Input
-                  required
-                  label="Giá gốc khóa học"
-                  placeholder="1.200.000 đ"
-                  errorMessage={
-                    formik.touched.originalPrice
-                      ? formik.errors.originalPrice
-                      : ""
-                  }
-                  {...formik.getFieldProps("originalPrice")}
-                />
-                <FormControl.Input
-                  required
-                  label="Giá khuyến mãi khóa học"
-                  placeholder="1.000.000 đ"
-                  errorMessage={
-                    formik.touched.currentPrice
-                      ? formik.errors.currentPrice
-                      : ""
-                  }
-                  {...formik.getFieldProps("currentPrice")}
-                />
-                <div className="editor">
-                  <h2>
-                    Nội dung khóa học <span>*</span>
-                  </h2>
-                  <FormControl.FormEditor
-                    value={formik.values.description}
-                    onChange={(value) =>
-                      formik.setFieldValue("description", value)
-                    }
-                    placeholder="Thêm một mô tả. Bao gồm những gì học sinh sẽ có thể làm sau khi hoàn thành bài giảng."
-                  />
-                </div>
-                {formik.touched.description && (
-                  <div className="editor-error">
-                    {formik.errors.description}
-                  </div>
-                )}
-                <FormControl.InputSelect
-                  label="Loại khóa học"
-                  list={categories}
-                  onChange={(category_type) =>
-                    formik.setFieldValue("category", category_type)
-                  }
-                  defaultValue={formik.values.category}
-                />
-              </Box>
-              <Box
-                sx={{
-                  width: "max-content",
-                  marginLeft: "auto",
-                }}
-              >
-                <Button
-                  type="submit"
-                  variant="contained"
-                  color="primary"
-                  sx={{
-                    height: 45,
-                  }}
-                >
-                  Lưu thông tin
-                </Button>
-              </Box>
-            </form>
+            <CourseDetailSideBarItem.CourseInformation
+              slug={slug}
+              courseStatus={courseStatus}
+              isUpdateCompleted={(status) =>
+                setIsInformationCourseUpdated(status)
+              }
+              courseInformationValue={formik.values}
+            />
           ) : navbar === 1 ? (
-            <div className="chapter-list">
-              {chapters.map((chapter, index) => (
-                <React.Fragment key={index}>
-                  <div className="new">
-                    <div
-                      className="icon"
-                      onClick={() => handleAddChapter(index + 1)}
-                    >
-                      <MediaContent.Icon icon="plus" color="black" size={20} />
-                    </div>
-                  </div>
-                  <Chapter
-                    chapter={chapter}
-                    index={index}
-                    handleUpdateChapter={handleUpdateChapter}
-                    handleDeleteChapter={handleDeleteChapter}
-                  />
-                </React.Fragment>
-              ))}
-              <div className="new">
-                <div
-                  className="icon"
-                  onClick={() => handleAddChapter(chapters.length + 1)}
-                >
-                  <MediaContent.Icon icon="plus" color="black" size={20} />
-                </div>
-              </div>
-            </div>
+            <CourseDetailSideBarItem.LearningContent
+              chapters={chapters}
+              isUpdateCompleted={(status) => setIsChapterUpdated(status)}
+            />
           ) : navbar === 2 ? (
-            <IntendedLearners />
+            <CourseDetailSideBarItem.IntendedLearners />
           ) : navbar === 3 ? (
-            <Requirements />
+            <CourseDetailSideBarItem.Requirements />
           ) : (
-            <Targets />
+            <CourseDetailSideBarItem.Targets />
           )}
         </div>
       </div>
